Divide YJIT architecture dependent codes
