+++ yjit/src/asm/x86_64/mod.rs
@@ -67,7 +67,7 @@ pub struct X86Mem
 }
 
 #[derive(Clone, Copy, Debug)]
-pub enum X86Opnd
+pub enum YJitOpnd
 {
     // Dummy operand
     None,
@@ -88,22 +88,22 @@ pub enum X86Opnd
     IPRel(i32)
 }
 
-impl X86Opnd {
+impl YJitOpnd {
     fn rex_needed(&self) -> bool {
         match self {
-            X86Opnd::None => false,
-            X86Opnd::Imm(_) => false,
-            X86Opnd::UImm(_) => false,
-            X86Opnd::Reg(reg) => reg.reg_no > 7 || reg.num_bits == 8 && reg.reg_no >= 4,
-            X86Opnd::Mem(mem) => (mem.base_reg_no > 7 || (mem.idx_reg_no.unwrap_or(0) > 7)),
-            X86Opnd::IPRel(_) => false
+            YJitOpnd::None => false,
+            YJitOpnd::Imm(_) => false,
+            YJitOpnd::UImm(_) => false,
+            YJitOpnd::Reg(reg) => reg.reg_no > 7 || reg.num_bits == 8 && reg.reg_no >= 4,
+            YJitOpnd::Mem(mem) => (mem.base_reg_no > 7 || (mem.idx_reg_no.unwrap_or(0) > 7)),
+            YJitOpnd::IPRel(_) => false
         }
     }
 
     // Check if an SIB byte is needed to encode this operand
     fn sib_needed(&self) -> bool {
         match self {
-            X86Opnd::Mem(mem) => {
+            YJitOpnd::Mem(mem) => {
                 mem.idx_reg_no.is_some() ||
                 mem.base_reg_no == RSP_REG_NO ||
                 mem.base_reg_no == R12_REG_NO
@@ -114,8 +114,8 @@ impl X86Opnd {
 
     fn disp_size(&self) -> u32 {
         match self {
-            X86Opnd::IPRel(_) => 32,
-            X86Opnd::Mem(mem) => {
+            YJitOpnd::IPRel(_) => 32,
+            YJitOpnd::Mem(mem) => {
                 if mem.disp != 0 {
                     // Compute the required displacement size
                     let num_bits = sig_imm_size(mem.disp.into());
@@ -138,17 +138,17 @@ impl X86Opnd {
 
     pub fn num_bits(&self) -> u8 {
         match self {
-            X86Opnd::Reg(reg) => reg.num_bits,
-            X86Opnd::Imm(imm) => imm.num_bits,
-            X86Opnd::UImm(uimm) => uimm.num_bits,
-            X86Opnd::Mem(mem) => mem.num_bits,
+            YJitOpnd::Reg(reg) => reg.num_bits,
+            YJitOpnd::Imm(imm) => imm.num_bits,
+            YJitOpnd::UImm(uimm) => uimm.num_bits,
+            YJitOpnd::Mem(mem) => mem.num_bits,
             _ => unreachable!()
         }
     }
 }
 
 // Instruction pointer
-pub const RIP: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::IP, reg_no: 5 });
+pub const RIP: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::IP, reg_no: 5 });
 
 // 64-bit GP registers
 const RAX_REG_NO: u8 = 0;
@@ -157,79 +157,79 @@ const RBP_REG_NO: u8 = 5;
 const R12_REG_NO: u8 = 12;
 const R13_REG_NO: u8 = 13;
 
-pub const RAX: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: RAX_REG_NO });
-pub const RCX: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 1 });
-pub const RDX: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 2 });
-pub const RBX: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 3 });
-pub const RSP: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: RSP_REG_NO });
-pub const RBP: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: RBP_REG_NO });
-pub const RSI: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 6 });
-pub const RDI: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 7 });
-pub const R8:  X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 8 });
-pub const R9:  X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 9 });
-pub const R10: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 10 });
-pub const R11: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 11 });
-pub const R12: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: R12_REG_NO });
-pub const R13: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: R13_REG_NO });
-pub const R14: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 14 });
-pub const R15: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 15 });
+pub const RAX: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: RAX_REG_NO });
+pub const RCX: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 1 });
+pub const RDX: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 2 });
+pub const RBX: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 3 });
+pub const RSP: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: RSP_REG_NO });
+pub const RBP: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: RBP_REG_NO });
+pub const RSI: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 6 });
+pub const RDI: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 7 });
+pub const R8:  YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 8 });
+pub const R9:  YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 9 });
+pub const R10: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 10 });
+pub const R11: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 11 });
+pub const R12: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: R12_REG_NO });
+pub const R13: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: R13_REG_NO });
+pub const R14: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 14 });
+pub const R15: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 64, reg_type: RegType::GP, reg_no: 15 });
 
 // 32-bit GP registers
-pub const EAX: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 0 });
-pub const ECX: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 1 });
-pub const EDX: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 2 });
-pub const EBX: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 3 });
-pub const ESP: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 4 });
-pub const EBP: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 5 });
-pub const ESI: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 6 });
-pub const EDI: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 7 });
-pub const R8D: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 8 });
-pub const R9D: X86Opnd  = X86Opnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 9 });
-pub const R10D: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 10 });
-pub const R11D: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 11 });
-pub const R12D: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 12 });
-pub const R13D: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 13 });
-pub const R14D: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 14 });
-pub const R15D: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 15 });
+pub const EAX: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 0 });
+pub const ECX: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 1 });
+pub const EDX: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 2 });
+pub const EBX: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 3 });
+pub const ESP: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 4 });
+pub const EBP: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 5 });
+pub const ESI: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 6 });
+pub const EDI: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 7 });
+pub const R8D: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 8 });
+pub const R9D: YJitOpnd  = YJitOpnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 9 });
+pub const R10D: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 10 });
+pub const R11D: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 11 });
+pub const R12D: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 12 });
+pub const R13D: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 13 });
+pub const R14D: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 14 });
+pub const R15D: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 32, reg_type: RegType::GP, reg_no: 15 });
 
 // 16-bit GP registers
-pub const AX:   X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 0 });
-pub const CX:   X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 1 });
-pub const DX:   X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 2 });
-pub const BX:   X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 3 });
-pub const SP:   X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 4 });
-pub const BP:   X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 5 });
-pub const SI:   X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 6 });
-pub const DI:   X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 7 });
-pub const R8W:  X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 8 });
-pub const R9W:  X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 9 });
-pub const R10W: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 10 });
-pub const R11W: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 11 });
-pub const R12W: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 12 });
-pub const R13W: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 13 });
-pub const R14W: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 14 });
-pub const R15W: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 15 });
+pub const AX:   YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 0 });
+pub const CX:   YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 1 });
+pub const DX:   YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 2 });
+pub const BX:   YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 3 });
+pub const SP:   YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 4 });
+pub const BP:   YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 5 });
+pub const SI:   YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 6 });
+pub const DI:   YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 7 });
+pub const R8W:  YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 8 });
+pub const R9W:  YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 9 });
+pub const R10W: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 10 });
+pub const R11W: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 11 });
+pub const R12W: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 12 });
+pub const R13W: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 13 });
+pub const R14W: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 14 });
+pub const R15W: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 16, reg_type: RegType::GP, reg_no: 15 });
 
 // 8-bit GP registers
-pub const AL:   X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 0 });
-pub const CL:   X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 1 });
-pub const DL:   X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 2 });
-pub const BL:   X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 3 });
-pub const SPL:  X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 4 });
-pub const BPL:  X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 5 });
-pub const SIL:  X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 6 });
-pub const DIL:  X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 7 });
-pub const R8B:  X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 8 });
-pub const R9B:  X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 9 });
-pub const R10B: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 10 });
-pub const R11B: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 11 });
-pub const R12B: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 12 });
-pub const R13B: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 13 });
-pub const R14B: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 14 });
-pub const R15B: X86Opnd = X86Opnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 15 });
+pub const AL:   YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 0 });
+pub const CL:   YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 1 });
+pub const DL:   YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 2 });
+pub const BL:   YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 3 });
+pub const SPL:  YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 4 });
+pub const BPL:  YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 5 });
+pub const SIL:  YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 6 });
+pub const DIL:  YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 7 });
+pub const R8B:  YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 8 });
+pub const R9B:  YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 9 });
+pub const R10B: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 10 });
+pub const R11B: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 11 });
+pub const R12B: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 12 });
+pub const R13B: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 13 });
+pub const R14B: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 14 });
+pub const R15B: YJitOpnd = YJitOpnd::Reg(X86Reg { num_bits: 8, reg_type: RegType::GP, reg_no: 15 });
 
 // C argument registers
-pub const C_ARG_REGS: [X86Opnd; 6] = [RDI, RSI, RDX, RCX, R8, R9];
+pub const C_ARG_REGS: [YJitOpnd; 6] = [RDI, RSI, RDX, RCX, R8, R9];
 
 //===========================================================================
 
@@ -268,17 +268,17 @@ pub fn unsig_imm_size(imm: u64) -> u8
 }
 
 /// Shorthand for memory operand with base register and displacement
-pub fn mem_opnd(num_bits: u8, base_reg: X86Opnd, disp: i32) -> X86Opnd
+pub fn mem_opnd(num_bits: u8, base_reg: YJitOpnd, disp: i32) -> YJitOpnd
 {
     let base_reg = match base_reg {
-        X86Opnd::Reg(reg) => reg,
+        YJitOpnd::Reg(reg) => reg,
         _ => unreachable!()
     };
 
     if base_reg.reg_type == RegType::IP {
-        X86Opnd::IPRel(disp)
+        YJitOpnd::IPRel(disp)
     } else {
-        X86Opnd::Mem(
+        YJitOpnd::Mem(
             X86Mem {
                 num_bits: num_bits,
                 base_reg_no: base_reg.reg_no,
@@ -291,8 +291,8 @@ pub fn mem_opnd(num_bits: u8, base_reg: X86Opnd, disp: i32) -> X86Opnd
 }
 
 /// Memory operand with SIB (Scale Index Base) indexing
-pub fn mem_opnd_sib(num_bits: u8, base_opnd: X86Opnd, index_opnd: X86Opnd, scale: i32, disp: i32) -> X86Opnd {
-    if let (X86Opnd::Reg(base_reg), X86Opnd::Reg(index_reg)) = (base_opnd, index_opnd) {
+pub fn mem_opnd_sib(num_bits: u8, base_opnd: YJitOpnd, index_opnd: YJitOpnd, scale: i32, disp: i32) -> YJitOpnd {
+    if let (YJitOpnd::Reg(base_reg), YJitOpnd::Reg(index_reg)) = (base_opnd, index_opnd) {
         let scale_exp: u8;
 
         match scale {
@@ -303,7 +303,7 @@ pub fn mem_opnd_sib(num_bits: u8, base_opnd: X86Opnd, index_opnd: X86Opnd, scale
             _ => unreachable!()
         };
 
-        X86Opnd::Mem(X86Mem {
+        YJitOpnd::Mem(X86Mem {
             num_bits,
             base_reg_no: base_reg.reg_no,
             idx_reg_no: Some(index_reg.reg_no),
@@ -333,7 +333,7 @@ pub fn mem_opnd_sib(num_bits: u8, base_opnd: X86Opnd, index_opnd: X86Opnd, scale
 */
 
 /*
-// TODO: this should be a method, X86Opnd.resize() or X86Opnd.subreg()
+// TODO: this should be a method, YJitOpnd.resize() or YJitOpnd.subreg()
 static x86opnd_t resize_opnd(x86opnd_t opnd, uint32_t num_bits)
 {
     assert (num_bits % 8 == 0);
@@ -343,22 +343,22 @@ static x86opnd_t resize_opnd(x86opnd_t opnd, uint32_t num_bits)
 }
 */
 
-pub fn imm_opnd(value: i64) -> X86Opnd
+pub fn imm_opnd(value: i64) -> YJitOpnd
 {
-    X86Opnd::Imm(X86Imm { num_bits: sig_imm_size(value), value })
+    YJitOpnd::Imm(X86Imm { num_bits: sig_imm_size(value), value })
 }
 
-pub fn uimm_opnd(value: u64) -> X86Opnd
+pub fn uimm_opnd(value: u64) -> YJitOpnd
 {
-    X86Opnd::UImm(X86UImm { num_bits: unsig_imm_size(value), value })
+    YJitOpnd::UImm(X86UImm { num_bits: unsig_imm_size(value), value })
 }
 
-pub fn const_ptr_opnd(ptr: *const u8) -> X86Opnd
+pub fn const_ptr_opnd(ptr: *const u8) -> YJitOpnd
 {
     uimm_opnd(ptr as u64)
 }
 
-pub fn code_ptr_opnd(code_ptr: CodePtr) -> X86Opnd
+pub fn code_ptr_opnd(code_ptr: CodePtr) -> YJitOpnd
 {
     uimm_opnd(code_ptr.raw_ptr() as u64)
 }
@@ -386,10 +386,10 @@ fn write_opcode(cb: &mut CodeBlock, opcode: u8, reg: X86Reg) {
 }
 
 /// Encode an RM instruction
-fn write_rm(cb: &mut CodeBlock, sz_pref: bool, rex_w: bool, r_opnd: X86Opnd, rm_opnd: X86Opnd, op_ext: u8, bytes: &[u8]) {
+fn write_rm(cb: &mut CodeBlock, sz_pref: bool, rex_w: bool, r_opnd: YJitOpnd, rm_opnd: YJitOpnd, op_ext: u8, bytes: &[u8]) {
     let op_len = bytes.len();
     assert!(op_len > 0 && op_len <= 3);
-    assert!(matches!(r_opnd, X86Opnd::Reg(_) | X86Opnd::None), "Can only encode an RM instruction with a register or a none");
+    assert!(matches!(r_opnd, YJitOpnd::Reg(_) | YJitOpnd::None), "Can only encode an RM instruction with a register or a none");
 
     // Flag to indicate the REX prefix is needed
     let need_rex = rex_w || r_opnd.rex_needed() || rm_opnd.rex_needed();
@@ -412,19 +412,19 @@ fn write_rm(cb: &mut CodeBlock, sz_pref: bool, rex_w: bool, r_opnd: X86Opnd, rm_
 
         let w = if rex_w { 1 } else { 0 };
         let r = match r_opnd {
-            X86Opnd::None => 0,
-            X86Opnd::Reg(reg) => if (reg.reg_no & 8) > 0 { 1 } else { 0 },
+            YJitOpnd::None => 0,
+            YJitOpnd::Reg(reg) => if (reg.reg_no & 8) > 0 { 1 } else { 0 },
             _ => unreachable!()
         };
 
         let x = match (need_sib, rm_opnd) {
-            (true, X86Opnd::Mem(mem)) => if (mem.idx_reg_no.unwrap_or(0) & 8) > 0 { 1 } else { 0 },
+            (true, YJitOpnd::Mem(mem)) => if (mem.idx_reg_no.unwrap_or(0) & 8) > 0 { 1 } else { 0 },
             _ => 0
         };
 
         let b = match rm_opnd {
-            X86Opnd::Reg(reg) => if (reg.reg_no & 8) > 0 { 1 } else { 0 },
-            X86Opnd::Mem(mem) => if (mem.base_reg_no & 8) > 0 { 1 } else { 0 },
+            YJitOpnd::Reg(reg) => if (reg.reg_no & 8) > 0 { 1 } else { 0 },
+            YJitOpnd::Mem(mem) => if (mem.base_reg_no & 8) > 0 { 1 } else { 0 },
             _ => 0
         };
 
@@ -443,15 +443,15 @@ fn write_rm(cb: &mut CodeBlock, sz_pref: bool, rex_w: bool, r_opnd: X86Opnd, rm_
     // MODRM.rm  (3 bits)
 
     assert!(
-        !(op_ext != 0xff && !matches!(r_opnd, X86Opnd::None)),
+        !(op_ext != 0xff && !matches!(r_opnd, YJitOpnd::None)),
         "opcode extension and register operand present"
     );
 
     // Encode the mod field
     let rm_mod = match rm_opnd {
-        X86Opnd::Reg(_) => 3,
-        X86Opnd::IPRel(_) => 0,
-        X86Opnd::Mem(_mem) => {
+        YJitOpnd::Reg(_) => 3,
+        YJitOpnd::IPRel(_) => 0,
+        YJitOpnd::Mem(_mem) => {
             match rm_opnd.disp_size() {
                 0 => 0,
                 8 => 1,
@@ -468,16 +468,16 @@ fn write_rm(cb: &mut CodeBlock, sz_pref: bool, rex_w: bool, r_opnd: X86Opnd, rm_
         reg = op_ext;
     } else {
         reg = match r_opnd {
-            X86Opnd::Reg(reg) => reg.reg_no & 7,
+            YJitOpnd::Reg(reg) => reg.reg_no & 7,
             _ => 0
         };
     }
 
     // Encode the rm field
     let rm = match rm_opnd {
-        X86Opnd::Reg(reg) => reg.reg_no & 7,
-        X86Opnd::Mem(mem) => if need_sib { 4 } else { mem.base_reg_no & 7 },
-        X86Opnd::IPRel(_) => 0b101,
+        YJitOpnd::Reg(reg) => reg.reg_no & 7,
+        YJitOpnd::Mem(mem) => if need_sib { 4 } else { mem.base_reg_no & 7 },
+        YJitOpnd::IPRel(_) => 0b101,
         _ => unreachable!()
     };
 
@@ -492,7 +492,7 @@ fn write_rm(cb: &mut CodeBlock, sz_pref: bool, rex_w: bool, r_opnd: X86Opnd, rm_
         // SIB.base  (3 bits)
 
         match rm_opnd {
-            X86Opnd::Mem(mem) => {
+            YJitOpnd::Mem(mem) => {
                 // Encode the scale value
                 let scale = mem.scale_exp;
 
@@ -512,13 +512,13 @@ fn write_rm(cb: &mut CodeBlock, sz_pref: bool, rex_w: bool, r_opnd: X86Opnd, rm_
 
     // Add the displacement
     match rm_opnd {
-        X86Opnd::Mem(mem) => {
+        YJitOpnd::Mem(mem) => {
             let disp_size = rm_opnd.disp_size();
             if disp_size > 0 {
                 cb.write_int(mem.disp as u64, disp_size);
             }
         },
-        X86Opnd::IPRel(rel) => {
+        YJitOpnd::IPRel(rel) => {
             cb.write_int(rel as u64, 32);
         },
         _ => ()
@@ -526,24 +526,24 @@ fn write_rm(cb: &mut CodeBlock, sz_pref: bool, rex_w: bool, r_opnd: X86Opnd, rm_
 }
 
 // Encode a mul-like single-operand RM instruction
-fn write_rm_unary(cb: &mut CodeBlock, op_mem_reg_8: u8, op_mem_reg_pref: u8, op_ext: u8, opnd: X86Opnd) {
-    assert!(matches!(opnd, X86Opnd::Reg(_) | X86Opnd::Mem(_)));
+fn write_rm_unary(cb: &mut CodeBlock, op_mem_reg_8: u8, op_mem_reg_pref: u8, op_ext: u8, opnd: YJitOpnd) {
+    assert!(matches!(opnd, YJitOpnd::Reg(_) | YJitOpnd::Mem(_)));
 
     let opnd_size = opnd.num_bits();
     assert!(opnd_size == 8 || opnd_size == 16 || opnd_size == 32 || opnd_size == 64);
 
     if opnd_size == 8 {
-        write_rm(cb, false, false, X86Opnd::None, opnd, op_ext, &[op_mem_reg_8]);
+        write_rm(cb, false, false, YJitOpnd::None, opnd, op_ext, &[op_mem_reg_8]);
     } else {
         let sz_pref = opnd_size == 16;
         let rex_w = opnd_size == 64;
-        write_rm(cb, sz_pref, rex_w, X86Opnd::None, opnd, op_ext, &[op_mem_reg_pref]);
+        write_rm(cb, sz_pref, rex_w, YJitOpnd::None, opnd, op_ext, &[op_mem_reg_pref]);
     }
 }
 
 // Encode an add-like RM instruction with multiple possible encodings
-fn write_rm_multi(cb: &mut CodeBlock, op_mem_reg8: u8, op_mem_reg_pref: u8, op_reg_mem8: u8, op_reg_mem_pref: u8, op_mem_imm8: u8, op_mem_imm_sml: u8, op_mem_imm_lrg: u8, op_ext_imm: u8, opnd0: X86Opnd, opnd1: X86Opnd) {
-    assert!(matches!(opnd0, X86Opnd::Reg(_) | X86Opnd::Mem(_)));
+fn write_rm_multi(cb: &mut CodeBlock, op_mem_reg8: u8, op_mem_reg_pref: u8, op_reg_mem8: u8, op_reg_mem_pref: u8, op_mem_imm8: u8, op_mem_imm_sml: u8, op_mem_imm_lrg: u8, op_ext_imm: u8, opnd0: YJitOpnd, opnd1: YJitOpnd) {
+    assert!(matches!(opnd0, YJitOpnd::Reg(_) | YJitOpnd::Mem(_)));
 
     // Check the size of opnd0
     let opnd_size = opnd0.num_bits();
@@ -551,10 +551,10 @@ fn write_rm_multi(cb: &mut CodeBlock, op_mem_reg8: u8, op_mem_reg_pref: u8, op_r
 
     // Check the size of opnd1
     match opnd1 {
-        X86Opnd::Reg(reg) => assert!(reg.num_bits == opnd_size),
-        X86Opnd::Mem(mem) => assert!(mem.num_bits == opnd_size),
-        X86Opnd::Imm(imm) => assert!(imm.num_bits <= opnd_size),
-        X86Opnd::UImm(uimm) => assert!(uimm.num_bits <= opnd_size),
+        YJitOpnd::Reg(reg) => assert!(reg.num_bits == opnd_size),
+        YJitOpnd::Mem(mem) => assert!(mem.num_bits == opnd_size),
+        YJitOpnd::Imm(imm) => assert!(imm.num_bits <= opnd_size),
+        YJitOpnd::UImm(uimm) => assert!(uimm.num_bits <= opnd_size),
         _ => ()
     };
 
@@ -563,7 +563,7 @@ fn write_rm_multi(cb: &mut CodeBlock, op_mem_reg8: u8, op_mem_reg_pref: u8, op_r
 
     match (opnd0, opnd1) {
         // R/M + Reg
-        (X86Opnd::Mem(_), X86Opnd::Reg(_)) | (X86Opnd::Reg(_), X86Opnd::Reg(_)) => {
+        (YJitOpnd::Mem(_), YJitOpnd::Reg(_)) | (YJitOpnd::Reg(_), YJitOpnd::Reg(_)) => {
             if opnd_size == 8 {
                 write_rm(cb, false, false, opnd1, opnd0, 0xff, &[op_mem_reg8]);
             } else {
@@ -571,7 +571,7 @@ fn write_rm_multi(cb: &mut CodeBlock, op_mem_reg8: u8, op_mem_reg_pref: u8, op_r
             }
         },
         // Reg + R/M/IPRel
-        (X86Opnd::Reg(_), X86Opnd::Mem(_) | X86Opnd::IPRel(_)) => {
+        (YJitOpnd::Reg(_), YJitOpnd::Mem(_) | YJitOpnd::IPRel(_)) => {
             if opnd_size == 8 {
                 write_rm(cb, false, false, opnd0, opnd1, 0xff, &[op_reg_mem8]);
             } else {
@@ -579,14 +579,14 @@ fn write_rm_multi(cb: &mut CodeBlock, op_mem_reg8: u8, op_mem_reg_pref: u8, op_r
             }
         },
         // R/M + Imm
-        (_, X86Opnd::Imm(imm)) => {
+        (_, YJitOpnd::Imm(imm)) => {
             if imm.num_bits <= 8 {
                 // 8-bit immediate
 
                 if opnd_size == 8 {
-                    write_rm(cb, false, false, X86Opnd::None, opnd0, op_ext_imm, &[op_mem_imm8]);
+                    write_rm(cb, false, false, YJitOpnd::None, opnd0, op_ext_imm, &[op_mem_imm8]);
                 } else {
-                    write_rm(cb, sz_pref, rex_w, X86Opnd::None, opnd0, op_ext_imm, &[op_mem_imm_sml]);
+                    write_rm(cb, sz_pref, rex_w, YJitOpnd::None, opnd0, op_ext_imm, &[op_mem_imm_sml]);
                 }
 
                 cb.write_int(imm.value as u64, 8);
@@ -594,23 +594,23 @@ fn write_rm_multi(cb: &mut CodeBlock, op_mem_reg8: u8, op_mem_reg_pref: u8, op_r
                 // 32-bit immediate
 
                 assert!(imm.num_bits <= opnd_size);
-                write_rm(cb, sz_pref, rex_w, X86Opnd::None, opnd0, op_ext_imm, &[op_mem_imm_lrg]);
+                write_rm(cb, sz_pref, rex_w, YJitOpnd::None, opnd0, op_ext_imm, &[op_mem_imm_lrg]);
                 cb.write_int(imm.value as u64, if opnd_size > 32 { 32 } else { opnd_size.into() });
             } else {
                 panic!("immediate value too large");
             }
         },
         // R/M + UImm
-        (_, X86Opnd::UImm(uimm)) => {
+        (_, YJitOpnd::UImm(uimm)) => {
             let num_bits = sig_imm_size(uimm.value.try_into().unwrap());
 
             if num_bits <= 8 {
                 // 8-bit immediate
 
                 if opnd_size == 8 {
-                    write_rm(cb, false, false, X86Opnd::None, opnd0, op_ext_imm, &[op_mem_imm8]);
+                    write_rm(cb, false, false, YJitOpnd::None, opnd0, op_ext_imm, &[op_mem_imm8]);
                 } else {
-                    write_rm(cb, sz_pref, rex_w, X86Opnd::None, opnd0, op_ext_imm, &[op_mem_imm_sml]);
+                    write_rm(cb, sz_pref, rex_w, YJitOpnd::None, opnd0, op_ext_imm, &[op_mem_imm_sml]);
                 }
 
                 cb.write_int(uimm.value, 8);
@@ -618,7 +618,7 @@ fn write_rm_multi(cb: &mut CodeBlock, op_mem_reg8: u8, op_mem_reg_pref: u8, op_r
                 // 32-bit immediate
 
                 assert!(num_bits <= opnd_size);
-                write_rm(cb, sz_pref, rex_w, X86Opnd::None, opnd0, op_ext_imm, &[op_mem_imm_lrg]);
+                write_rm(cb, sz_pref, rex_w, YJitOpnd::None, opnd0, op_ext_imm, &[op_mem_imm_lrg]);
                 cb.write_int(uimm.value, if opnd_size > 32 { 32 } else { opnd_size.into() });
             } else {
                 panic!("immediate value too large");
@@ -634,7 +634,7 @@ pub fn write_lock_prefix(cb: &mut CodeBlock) {
 }
 
 /// add - Integer addition
-pub fn add(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
+pub fn add(cb: &mut CodeBlock, opnd0: YJitOpnd, opnd1: YJitOpnd) {
     write_rm_multi(
         cb,
         0x00, // opMemReg8
@@ -651,7 +651,7 @@ pub fn add(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
 }
 
 /// and - Bitwise AND
-pub fn and(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
+pub fn and(cb: &mut CodeBlock, opnd0: YJitOpnd, opnd1: YJitOpnd) {
     write_rm_multi(
         cb,
         0x20, // opMemReg8
@@ -677,8 +677,8 @@ pub fn call_rel32(cb: &mut CodeBlock, rel32: i32) {
 }
 
 /// call - Call a pointer, encode with a 32-bit offset if possible
-pub fn call_ptr(cb: &mut CodeBlock, scratch_opnd: X86Opnd, dst_ptr: *const u8) {
-    if let X86Opnd::Reg(_scratch_reg) = scratch_opnd {
+pub fn call_ptr(cb: &mut CodeBlock, scratch_opnd: YJitOpnd, dst_ptr: *const u8) {
+    if let YJitOpnd::Reg(_scratch_reg) = scratch_opnd {
         // Pointer to the end of this call instruction
         let end_ptr = cb.get_ptr(cb.write_pos + 5);
 
@@ -712,16 +712,16 @@ pub fn call_label(cb: &mut CodeBlock, label_idx: usize) {
 }
 
 /// call - Indirect call with an R/M operand
-pub fn call(cb: &mut CodeBlock, opnd: X86Opnd) {
-    write_rm(cb, false, false, X86Opnd::None, opnd, 2, &[0xff]);
+pub fn call(cb: &mut CodeBlock, opnd: YJitOpnd) {
+    write_rm(cb, false, false, YJitOpnd::None, opnd, 2, &[0xff]);
 }
 
 /// Encode a conditional move instruction
-fn write_cmov(cb: &mut CodeBlock, opcode1: u8, dst: X86Opnd, src: X86Opnd) {
-    if let X86Opnd::Reg(reg) = dst {
+fn write_cmov(cb: &mut CodeBlock, opcode1: u8, dst: YJitOpnd, src: YJitOpnd) {
+    if let YJitOpnd::Reg(reg) = dst {
         match src {
-            X86Opnd::Reg(_) => (),
-            X86Opnd::Mem(_) => (),
+            YJitOpnd::Reg(_) => (),
+            YJitOpnd::Mem(_) => (),
             _ => unreachable!()
         };
 
@@ -736,39 +736,39 @@ fn write_cmov(cb: &mut CodeBlock, opcode1: u8, dst: X86Opnd, src: X86Opnd) {
 }
 
 // cmovcc - Conditional move
-pub fn cmova(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x47, dst, src); }
-pub fn cmovae(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x43, dst, src); }
-pub fn cmovb(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x42, dst, src); }
-pub fn cmovbe(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x46, dst, src); }
-pub fn cmovc(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x42, dst, src); }
-pub fn cmove(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x44, dst, src); }
-pub fn cmovg(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x4f, dst, src); }
-pub fn cmovge(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x4d, dst, src); }
-pub fn cmovl(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x4c, dst, src); }
-pub fn cmovle(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x4e, dst, src); }
-pub fn cmovna(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x46, dst, src); }
-pub fn cmovnae(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x42, dst, src); }
-pub fn cmovnb(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x43, dst, src); }
-pub fn cmovnbe(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x47, dst, src); }
-pub fn cmovnc(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x43, dst, src); }
-pub fn cmovne(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x45, dst, src); }
-pub fn cmovng(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x4e, dst, src); }
-pub fn cmovnge(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x4c, dst, src); }
-pub fn cmovnl(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb,  0x4d, dst, src); }
-pub fn cmovnle(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x4f, dst, src); }
-pub fn cmovno(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x41, dst, src); }
-pub fn cmovnp(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x4b, dst, src); }
-pub fn cmovns(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x49, dst, src); }
-pub fn cmovnz(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x45, dst, src); }
-pub fn cmovo(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x40, dst, src); }
-pub fn cmovp(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x4a, dst, src); }
-pub fn cmovpe(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x4a, dst, src); }
-pub fn cmovpo(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x4b, dst, src); }
-pub fn cmovs(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x48, dst, src); }
-pub fn cmovz(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) { write_cmov(cb, 0x44, dst, src); }
+pub fn cmova(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x47, dst, src); }
+pub fn cmovae(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x43, dst, src); }
+pub fn cmovb(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x42, dst, src); }
+pub fn cmovbe(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x46, dst, src); }
+pub fn cmovc(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x42, dst, src); }
+pub fn cmove(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x44, dst, src); }
+pub fn cmovg(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x4f, dst, src); }
+pub fn cmovge(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x4d, dst, src); }
+pub fn cmovl(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x4c, dst, src); }
+pub fn cmovle(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x4e, dst, src); }
+pub fn cmovna(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x46, dst, src); }
+pub fn cmovnae(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x42, dst, src); }
+pub fn cmovnb(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x43, dst, src); }
+pub fn cmovnbe(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x47, dst, src); }
+pub fn cmovnc(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x43, dst, src); }
+pub fn cmovne(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x45, dst, src); }
+pub fn cmovng(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x4e, dst, src); }
+pub fn cmovnge(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x4c, dst, src); }
+pub fn cmovnl(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb,  0x4d, dst, src); }
+pub fn cmovnle(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x4f, dst, src); }
+pub fn cmovno(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x41, dst, src); }
+pub fn cmovnp(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x4b, dst, src); }
+pub fn cmovns(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x49, dst, src); }
+pub fn cmovnz(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x45, dst, src); }
+pub fn cmovo(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x40, dst, src); }
+pub fn cmovp(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x4a, dst, src); }
+pub fn cmovpe(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x4a, dst, src); }
+pub fn cmovpo(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x4b, dst, src); }
+pub fn cmovs(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x48, dst, src); }
+pub fn cmovz(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) { write_cmov(cb, 0x44, dst, src); }
 
 /// cmp - Compare and set flags
-pub fn cmp(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
+pub fn cmp(cb: &mut CodeBlock, opnd0: YJitOpnd, opnd1: YJitOpnd) {
     write_rm_multi(
         cb,
         0x38, // opMemReg8
@@ -907,9 +907,13 @@ pub fn js_ptr  (cb: &mut CodeBlock, ptr: CodePtr) { write_jcc_ptr(cb, 0x0F, 0x88
 pub fn jz_ptr  (cb: &mut CodeBlock, ptr: CodePtr) { write_jcc_ptr(cb, 0x0F, 0x84, ptr); }
 pub fn jmp_ptr (cb: &mut CodeBlock, ptr: CodePtr) { write_jcc_ptr(cb, 0xFF, 0xE9, ptr); }
 
+pub fn gen_jump_ptr(cb: &mut CodeBlock, ptr: CodePtr) {
+    jmp_ptr(cb, ptr);
+}
+
 /// jmp - Indirect jump near to an R/M operand.
-pub fn jmp_rm(cb: &mut CodeBlock, opnd: X86Opnd) {
-    write_rm(cb, false, false, X86Opnd::None, opnd, 4, &[0xff]);
+pub fn jmp_rm(cb: &mut CodeBlock, opnd: YJitOpnd) {
+    write_rm(cb, false, false, YJitOpnd::None, opnd, 4, &[0xff]);
 }
 
 // jmp - Jump with relative 32-bit offset
@@ -919,8 +923,8 @@ pub fn jmp32(cb: &mut CodeBlock, offset: i32) {
 }
 
 /// lea - Load Effective Address
-pub fn lea(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) {
-    if let X86Opnd::Reg(reg) = dst {
+pub fn lea(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) {
+    if let YJitOpnd::Reg(reg) = dst {
         assert!(reg.num_bits == 64);
         write_rm(cb, false, true, dst, src, 0xff, &[0x8d]);
     } else {
@@ -929,10 +933,10 @@ pub fn lea(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) {
 }
 
 /// mov - Data move operation
-pub fn mov(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) {
+pub fn mov(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) {
     match (dst, src) {
         // R + Imm
-        (X86Opnd::Reg(reg), X86Opnd::Imm(imm)) => {
+        (YJitOpnd::Reg(reg), YJitOpnd::Imm(imm)) => {
             assert!(imm.num_bits <= reg.num_bits);
 
             // In case the source immediate could be zero extended to be 64
@@ -959,7 +963,7 @@ pub fn mov(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) {
             }
         },
         // R + UImm
-        (X86Opnd::Reg(reg), X86Opnd::UImm(uimm)) => {
+        (YJitOpnd::Reg(reg), YJitOpnd::UImm(uimm)) => {
             assert!(uimm.num_bits <= reg.num_bits);
 
             // In case the source immediate could be zero extended to be 64
@@ -986,13 +990,13 @@ pub fn mov(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) {
             }
         },
         // M + Imm
-        (X86Opnd::Mem(mem), X86Opnd::Imm(imm)) => {
+        (YJitOpnd::Mem(mem), YJitOpnd::Imm(imm)) => {
             assert!(imm.num_bits <= mem.num_bits);
 
             if mem.num_bits == 8 {
-                write_rm(cb, false, false, X86Opnd::None, dst, 0xff, &[0xc6]);
+                write_rm(cb, false, false, YJitOpnd::None, dst, 0xff, &[0xc6]);
             } else {
-                write_rm(cb, mem.num_bits == 16, mem.num_bits == 64, X86Opnd::None, dst, 0, &[0xc7]);
+                write_rm(cb, mem.num_bits == 16, mem.num_bits == 64, YJitOpnd::None, dst, 0, &[0xc7]);
             }
 
             let output_num_bits:u32 = if mem.num_bits > 32 { 32 } else { mem.num_bits.into() };
@@ -1000,14 +1004,14 @@ pub fn mov(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) {
             cb.write_int(imm.value as u64, output_num_bits);
         },
         // M + UImm
-        (X86Opnd::Mem(mem), X86Opnd::UImm(uimm)) => {
+        (YJitOpnd::Mem(mem), YJitOpnd::UImm(uimm)) => {
             assert!(uimm.num_bits <= mem.num_bits);
 
             if mem.num_bits == 8 {
-                write_rm(cb, false, false, X86Opnd::None, dst, 0xff, &[0xc6]);
+                write_rm(cb, false, false, YJitOpnd::None, dst, 0xff, &[0xc6]);
             }
             else {
-                write_rm(cb, mem.num_bits == 16, mem.num_bits == 64, X86Opnd::None, dst, 0, &[0xc7]);
+                write_rm(cb, mem.num_bits == 16, mem.num_bits == 64, YJitOpnd::None, dst, 0, &[0xc7]);
             }
 
             let output_num_bits = if mem.num_bits > 32 { 32 } else { mem.num_bits.into() };
@@ -1015,7 +1019,7 @@ pub fn mov(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) {
             cb.write_int(uimm.value, output_num_bits);
         },
         // * + Imm/UImm
-        (_, X86Opnd::Imm(_) | X86Opnd::UImm(_)) => unreachable!(),
+        (_, YJitOpnd::Imm(_) | YJitOpnd::UImm(_)) => unreachable!(),
         // * + *
         (_, _) => {
             write_rm_multi(
@@ -1036,9 +1040,9 @@ pub fn mov(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) {
 }
 
 /// movsx - Move with sign extension (signed integers)
-pub fn movsx(cb: &mut CodeBlock, dst: X86Opnd, src: X86Opnd) {
-    if let X86Opnd::Reg(_dst_reg) = dst {
-        assert!(matches!(src, X86Opnd::Reg(_) | X86Opnd::Mem(_)));
+pub fn movsx(cb: &mut CodeBlock, dst: YJitOpnd, src: YJitOpnd) {
+    if let YJitOpnd::Reg(_dst_reg) = dst {
+        assert!(matches!(src, YJitOpnd::Reg(_) | YJitOpnd::Mem(_)));
 
         let src_num_bits = src.num_bits();
         let dst_num_bits = dst.num_bits();
@@ -1120,7 +1124,7 @@ pub fn nop(cb: &mut CodeBlock, length: u32) {
 }
 
 /// not - Bitwise NOT
-pub fn not(cb: &mut CodeBlock, opnd: X86Opnd) {
+pub fn not(cb: &mut CodeBlock, opnd: YJitOpnd) {
     write_rm_unary(
         cb,
         0xf6, // opMemReg8
@@ -1131,7 +1135,7 @@ pub fn not(cb: &mut CodeBlock, opnd: X86Opnd) {
 }
 
 /// or - Bitwise OR
-pub fn or(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
+pub fn or(cb: &mut CodeBlock, opnd0: YJitOpnd, opnd1: YJitOpnd) {
     write_rm_multi(
         cb,
         0x08, // opMemReg8
@@ -1148,9 +1152,9 @@ pub fn or(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
 }
 
 /// pop - Pop a register off the stack
-pub fn pop(cb: &mut CodeBlock, opnd: X86Opnd) {
+pub fn pop(cb: &mut CodeBlock, opnd: YJitOpnd) {
     match opnd {
-        X86Opnd::Reg(reg) => {
+        YJitOpnd::Reg(reg) => {
             assert!(reg.num_bits == 64);
 
             if opnd.rex_needed() {
@@ -1158,10 +1162,10 @@ pub fn pop(cb: &mut CodeBlock, opnd: X86Opnd) {
             }
             write_opcode(cb, 0x58, reg);
         },
-        X86Opnd::Mem(mem) => {
+        YJitOpnd::Mem(mem) => {
             assert!(mem.num_bits == 64);
 
-            write_rm(cb, false, false, X86Opnd::None, opnd, 0, &[0x8f]);
+            write_rm(cb, false, false, YJitOpnd::None, opnd, 0, &[0x8f]);
         },
         _ => unreachable!()
     };
@@ -1174,16 +1178,16 @@ pub fn popfq(cb: &mut CodeBlock) {
 }
 
 /// push - Push an operand on the stack
-pub fn push(cb: &mut CodeBlock, opnd: X86Opnd) {
+pub fn push(cb: &mut CodeBlock, opnd: YJitOpnd) {
     match opnd {
-        X86Opnd::Reg(reg) => {
+        YJitOpnd::Reg(reg) => {
             if opnd.rex_needed() {
                 write_rex(cb, false, 0, 0, reg.reg_no);
             }
             write_opcode(cb, 0x50, reg);
         },
-        X86Opnd::Mem(_mem) => {
-            write_rm(cb, false, false, X86Opnd::None, opnd, 6, &[0xff]);
+        YJitOpnd::Mem(_mem) => {
+            write_rm(cb, false, false, YJitOpnd::None, opnd, 6, &[0xff]);
         },
         _ => unreachable!()
     }
@@ -1200,8 +1204,8 @@ pub fn ret(cb: &mut CodeBlock) {
 }
 
 // Encode a single-operand shift instruction
-fn write_shift(cb: &mut CodeBlock, op_mem_one_pref: u8, _op_mem_cl_pref: u8, op_mem_imm_pref: u8, op_ext: u8, opnd0: X86Opnd, opnd1: X86Opnd) {
-    assert!(matches!(opnd0, X86Opnd::Reg(_) | X86Opnd::Mem(_)));
+fn write_shift(cb: &mut CodeBlock, op_mem_one_pref: u8, _op_mem_cl_pref: u8, op_mem_imm_pref: u8, op_ext: u8, opnd0: YJitOpnd, opnd1: YJitOpnd) {
+    assert!(matches!(opnd0, YJitOpnd::Reg(_) | YJitOpnd::Mem(_)));
 
     // Check the size of opnd0
     let opnd_size = opnd0.num_bits();
@@ -1210,12 +1214,12 @@ fn write_shift(cb: &mut CodeBlock, op_mem_one_pref: u8, _op_mem_cl_pref: u8, op_
     let sz_pref = opnd_size == 16;
     let rex_w = opnd_size == 64;
 
-    if let X86Opnd::UImm(imm) = opnd1 {
+    if let YJitOpnd::UImm(imm) = opnd1 {
         if imm.value == 1 {
-            write_rm(cb, sz_pref, rex_w, X86Opnd::None, opnd0, op_ext, &[op_mem_one_pref]);
+            write_rm(cb, sz_pref, rex_w, YJitOpnd::None, opnd0, op_ext, &[op_mem_one_pref]);
         } else {
             assert!(imm.num_bits <= 8);
-            write_rm(cb, sz_pref, rex_w, X86Opnd::None, opnd0, op_ext, &[op_mem_imm_pref]);
+            write_rm(cb, sz_pref, rex_w, YJitOpnd::None, opnd0, op_ext, &[op_mem_imm_pref]);
             cb.write_byte(imm.value as u8);
         }
     } else {
@@ -1224,7 +1228,7 @@ fn write_shift(cb: &mut CodeBlock, op_mem_one_pref: u8, _op_mem_cl_pref: u8, op_
 }
 
 // sal - Shift arithmetic left
-pub fn sal(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
+pub fn sal(cb: &mut CodeBlock, opnd0: YJitOpnd, opnd1: YJitOpnd) {
     write_shift(
         cb,
         0xD1, // opMemOnePref,
@@ -1237,7 +1241,7 @@ pub fn sal(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
 }
 
 /// sar - Shift arithmetic right (signed)
-pub fn sar(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
+pub fn sar(cb: &mut CodeBlock, opnd0: YJitOpnd, opnd1: YJitOpnd) {
     write_shift(
         cb,
         0xD1, // opMemOnePref,
@@ -1250,7 +1254,7 @@ pub fn sar(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
 }
 
 // shl - Shift logical left
-pub fn shl(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
+pub fn shl(cb: &mut CodeBlock, opnd0: YJitOpnd, opnd1: YJitOpnd) {
     write_shift(
         cb,
         0xD1, // opMemOnePref,
@@ -1263,7 +1267,7 @@ pub fn shl(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
 }
 
 /// shr - Shift logical right (unsigned)
-pub fn shr(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
+pub fn shr(cb: &mut CodeBlock, opnd0: YJitOpnd, opnd1: YJitOpnd) {
     write_shift(
         cb,
         0xD1, // opMemOnePref,
@@ -1276,7 +1280,7 @@ pub fn shr(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
 }
 
 /// sub - Integer subtraction
-pub fn sub(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
+pub fn sub(cb: &mut CodeBlock, opnd0: YJitOpnd, opnd1: YJitOpnd) {
     write_rm_multi(
         cb,
         0x28, // opMemReg8
@@ -1292,29 +1296,29 @@ pub fn sub(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
     );
 }
 
-fn resize_opnd(opnd: X86Opnd, num_bits: u8) -> X86Opnd {
+fn resize_opnd(opnd: YJitOpnd, num_bits: u8) -> YJitOpnd {
     match opnd {
-        X86Opnd::Reg(reg) => {
+        YJitOpnd::Reg(reg) => {
             let mut cloned = reg;
             cloned.num_bits = num_bits;
-            X86Opnd::Reg(cloned)
+            YJitOpnd::Reg(cloned)
         },
-        X86Opnd::Mem(mem) => {
+        YJitOpnd::Mem(mem) => {
             let mut cloned = mem;
             cloned.num_bits = num_bits;
-            X86Opnd::Mem(cloned)
+            YJitOpnd::Mem(cloned)
         },
         _ => unreachable!()
     }
 }
 
 /// test - Logical Compare
-pub fn test(cb: &mut CodeBlock, rm_opnd: X86Opnd, test_opnd: X86Opnd) {
-    assert!(matches!(rm_opnd, X86Opnd::Reg(_) | X86Opnd::Mem(_)));
+pub fn test(cb: &mut CodeBlock, rm_opnd: YJitOpnd, test_opnd: YJitOpnd) {
+    assert!(matches!(rm_opnd, YJitOpnd::Reg(_) | YJitOpnd::Mem(_)));
     let rm_num_bits = rm_opnd.num_bits();
 
     match test_opnd {
-        X86Opnd::UImm(uimm) => {
+        YJitOpnd::UImm(uimm) => {
             assert!(uimm.num_bits <= 32);
             assert!(uimm.num_bits <= rm_num_bits);
 
@@ -1323,22 +1327,22 @@ pub fn test(cb: &mut CodeBlock, rm_opnd: X86Opnd, test_opnd: X86Opnd) {
             let rm_resized = resize_opnd(rm_opnd, uimm.num_bits);
 
             if uimm.num_bits == 8 {
-                write_rm(cb, false, false, X86Opnd::None, rm_resized, 0x00, &[0xf6]);
+                write_rm(cb, false, false, YJitOpnd::None, rm_resized, 0x00, &[0xf6]);
                 cb.write_int(uimm.value, uimm.num_bits.into());
             } else {
-                write_rm(cb, uimm.num_bits == 16, false, X86Opnd::None, rm_resized, 0x00, &[0xf7]);
+                write_rm(cb, uimm.num_bits == 16, false, YJitOpnd::None, rm_resized, 0x00, &[0xf7]);
                 cb.write_int(uimm.value, uimm.num_bits.into());
             }
         },
-        X86Opnd::Imm(imm) => {
+        YJitOpnd::Imm(imm) => {
             // This mode only applies to 64-bit R/M operands with 32-bit signed immediates
             assert!(imm.num_bits <= 32);
             assert!(rm_num_bits == 64);
 
-            write_rm(cb, false, true, X86Opnd::None, rm_opnd, 0x00, &[0xf7]);
+            write_rm(cb, false, true, YJitOpnd::None, rm_opnd, 0x00, &[0xf7]);
             cb.write_int(imm.value as u64, 32);
         },
-        X86Opnd::Reg(reg) => {
+        YJitOpnd::Reg(reg) => {
             assert!(reg.num_bits == rm_num_bits);
 
             if rm_num_bits == 8 {
@@ -1357,8 +1361,8 @@ pub fn ud2(cb: &mut CodeBlock) {
 }
 
 /// xchg - Exchange Register/Memory with Register
-pub fn xchg(cb: &mut CodeBlock, rm_opnd: X86Opnd, r_opnd: X86Opnd) {
-    if let (X86Opnd::Reg(rm_reg), X86Opnd::Reg(r_reg)) = (rm_opnd, r_opnd) {
+pub fn xchg(cb: &mut CodeBlock, rm_opnd: YJitOpnd, r_opnd: YJitOpnd) {
+    if let (YJitOpnd::Reg(rm_reg), YJitOpnd::Reg(r_reg)) = (rm_opnd, r_opnd) {
         assert!(rm_reg.num_bits == 64);
         assert!(r_reg.num_bits == 64);
 
@@ -1378,7 +1382,7 @@ pub fn xchg(cb: &mut CodeBlock, rm_opnd: X86Opnd, r_opnd: X86Opnd) {
 }
 
 /// xor - Exclusive bitwise OR
-pub fn xor(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) {
+pub fn xor(cb: &mut CodeBlock, opnd0: YJitOpnd, opnd1: YJitOpnd) {
     write_rm_multi(
         cb,
         0x30, // opMemReg8
