+++ yjit/src/codegen/x86_64.rs
@@ -1,120 +1,26 @@
-use crate::asm::x86_64::*;
 use crate::asm::*;
 use crate::core::*;
 use crate::cruby::*;
 use crate::invariants::*;
-use crate::options::*;
-use crate::stats::*;
 use crate::utils::*;
-use CodegenStatus::*;
-use InsnOpnd::*;
+use crate::stats::*;
+use crate::codegen::*;
 
-use std::cell::RefMut;
-use std::cmp;
-use std::collections::HashMap;
-use std::ffi::CStr;
-use std::mem::{self, size_of};
-use std::os::raw::c_uint;
+use std::mem::size_of;
 use std::ptr;
 use std::slice;
 
 // Callee-saved registers
-pub const REG_CFP: X86Opnd = R13;
-pub const REG_EC: X86Opnd = R12;
-pub const REG_SP: X86Opnd = RBX;
+pub const REG_CFP: YJitOpnd = R13;
+pub const REG_EC: YJitOpnd = R12;
+pub const REG_SP: YJitOpnd = RBX;
 
 // Scratch registers used by YJIT
-pub const REG0: X86Opnd = RAX;
-pub const REG0_32: X86Opnd = EAX;
-pub const REG0_8: X86Opnd = AL;
-pub const REG1: X86Opnd = RCX;
-// pub const REG1_32: X86Opnd = ECX;
-
-/// Status returned by code generation functions
-#[derive(PartialEq, Debug)]
-enum CodegenStatus {
-    EndBlock,
-    KeepCompiling,
-    CantCompile,
-}
-
-/// Code generation function signature
-type InsnGenFn = fn(
-    jit: &mut JITState,
-    ctx: &mut Context,
-    cb: &mut CodeBlock,
-    ocb: &mut OutlinedCb,
-) -> CodegenStatus;
-
-/// Code generation state
-/// This struct only lives while code is being generated
-pub struct JITState {
-    // Block version being compiled
-    block: BlockRef,
-
-    // Instruction sequence this is associated with
-    iseq: IseqPtr,
-
-    // Index of the current instruction being compiled
-    insn_idx: u32,
-
-    // Opcode for the instruction being compiled
-    opcode: usize,
-
-    // PC of the instruction being compiled
-    pc: *mut VALUE,
-
-    // Side exit to the instruction being compiled. See :side-exit:.
-    side_exit_for_pc: Option<CodePtr>,
-
-    // Execution context when compilation started
-    // This allows us to peek at run-time values
-    ec: Option<EcPtr>,
-
-    // Whether we need to record the code address at
-    // the end of this bytecode instruction for global invalidation
-    record_boundary_patch_point: bool,
-}
-
-impl JITState {
-    pub fn new(blockref: &BlockRef) -> Self {
-        JITState {
-            block: blockref.clone(),
-            iseq: ptr::null(), // TODO: initialize this from the blockid
-            insn_idx: 0,
-            opcode: 0,
-            pc: ptr::null_mut::<VALUE>(),
-            side_exit_for_pc: None,
-            ec: None,
-            record_boundary_patch_point: false,
-        }
-    }
-
-    pub fn get_block(&self) -> BlockRef {
-        self.block.clone()
-    }
-
-    pub fn get_insn_idx(&self) -> u32 {
-        self.insn_idx
-    }
-
-    pub fn get_iseq(self: &JITState) -> IseqPtr {
-        self.iseq
-    }
-
-    pub fn get_opcode(self: &JITState) -> usize {
-        self.opcode
-    }
-
-    pub fn add_gc_object_offset(self: &mut JITState, ptr_offset: u32) {
-        let mut gc_obj_vec: RefMut<_> = self.block.borrow_mut();
-        gc_obj_vec.add_gc_object_offset(ptr_offset);
-    }
-
-    pub fn get_pc(self: &JITState) -> *mut VALUE {
-        self.pc
-    }
-}
+pub const REG0: YJitOpnd = RAX;
+pub const REG0_32: YJitOpnd = EAX;
+pub const REG0_8: YJitOpnd = AL;
+pub const REG1: YJitOpnd = RCX;
+// pub const REG1_32: YJitOpnd = ECX;
 
 use crate::codegen::JCCKinds::*;
 
@@ -128,16 +34,9 @@ pub enum JCCKinds {
     JCC_JNA,
 }
 
-pub fn jit_get_arg(jit: &JITState, arg_idx: isize) -> VALUE {
-    // insn_len require non-test config
-    #[cfg(not(test))]
-    assert!(insn_len(jit.get_opcode()) > (arg_idx + 1).try_into().unwrap());
-    unsafe { *(jit.pc.offset(arg_idx + 1)) }
-}
-
 // Load a VALUE into a register and keep track of the reference if it is on the GC heap.
-pub fn jit_mov_gc_ptr(jit: &mut JITState, cb: &mut CodeBlock, reg: X86Opnd, ptr: VALUE) {
-    assert!(matches!(reg, X86Opnd::Reg(_)));
+pub fn jit_mov_gc_ptr(jit: &mut JITState, cb: &mut CodeBlock, reg: YJitOpnd, ptr: VALUE) {
+    assert!(matches!(reg, YJitOpnd::Reg(_)));
     assert!(reg.num_bits() == 64);
 
     // Load the pointer constant into the specified register
@@ -151,115 +50,17 @@ pub fn jit_mov_gc_ptr(jit: &mut JITState, cb: &mut CodeBlock, reg: X86Opnd, ptr:
     }
 }
 
-// Get the index of the next instruction
-fn jit_next_insn_idx(jit: &JITState) -> u32 {
-    jit.insn_idx + insn_len(jit.get_opcode())
-}
-
-// Check if we are compiling the instruction at the stub PC
-// Meaning we are compiling the instruction that is next to execute
-fn jit_at_current_insn(jit: &JITState) -> bool {
-    let ec_pc: *mut VALUE = unsafe { get_cfp_pc(get_ec_cfp(jit.ec.unwrap())) };
-    ec_pc == jit.pc
-}
-
-// Peek at the nth topmost value on the Ruby stack.
-// Returns the topmost value when n == 0.
-fn jit_peek_at_stack(jit: &JITState, ctx: &Context, n: isize) -> VALUE {
-    assert!(jit_at_current_insn(jit));
-    assert!(n < ctx.get_stack_size() as isize);
-
-    // Note: this does not account for ctx->sp_offset because
-    // this is only available when hitting a stub, and while
-    // hitting a stub, cfp->sp needs to be up to date in case
-    // codegen functions trigger GC. See :stub-sp-flush:.
-    return unsafe {
-        let sp: *mut VALUE = get_cfp_sp(get_ec_cfp(jit.ec.unwrap()));
-
-        *(sp.offset(-1 - n))
-    };
-}
-
-fn jit_peek_at_self(jit: &JITState) -> VALUE {
-    unsafe { get_cfp_self(get_ec_cfp(jit.ec.unwrap())) }
-}
-
-fn jit_peek_at_local(jit: &JITState, n: i32) -> VALUE {
-    assert!(jit_at_current_insn(jit));
-
-    let local_table_size: isize = unsafe { get_iseq_body_local_table_size(jit.iseq) }
-        .try_into()
-        .unwrap();
-    assert!(n < local_table_size.try_into().unwrap());
-
-    unsafe {
-        let ep = get_cfp_ep(get_ec_cfp(jit.ec.unwrap()));
-        let n_isize: isize = n.try_into().unwrap();
-        let offs: isize = -(VM_ENV_DATA_SIZE as isize) - local_table_size + n_isize + 1;
-        *ep.offset(offs)
-    }
-}
-
-// Add a comment at the current position in the code block
-fn add_comment(cb: &mut CodeBlock, comment_str: &str) {
-    if cfg!(feature = "asm_comments") {
-        cb.add_comment(comment_str);
-    }
-}
-
-/// Increment a profiling counter with counter_name
-#[cfg(not(feature = "stats"))]
-macro_rules! gen_counter_incr {
-    ($cb:tt, $counter_name:ident) => {};
-}
 #[cfg(feature = "stats")]
-macro_rules! gen_counter_incr {
-    ($cb:tt, $counter_name:ident) => {
-        if (get_option!(gen_stats)) {
-            // Get a pointer to the counter variable
-            let ptr = ptr_to_counter!($counter_name);
-
-            // Use REG1 because there might be return value in REG0
-            mov($cb, REG1, const_ptr_opnd(ptr as *const u8));
-            write_lock_prefix($cb); // for ractors.
-            add($cb, mem_opnd(64, REG1, 0), imm_opnd(1));
-        }
-    };
-}
-
-/// Increment a counter then take an existing side exit
-#[cfg(not(feature = "stats"))]
-macro_rules! counted_exit {
-    ($ocb:tt, $existing_side_exit:tt, $counter_name:ident) => {{
-        let _ = $ocb;
-        $existing_side_exit
-    }};
-}
-#[cfg(feature = "stats")]
-macro_rules! counted_exit {
-    ($ocb:tt, $existing_side_exit:tt, $counter_name:ident) => {
-        // The counter is only incremented when stats are enabled
-        if (!get_option!(gen_stats)) {
-            $existing_side_exit
-        } else {
-            let ocb = $ocb.unwrap();
-            let code_ptr = ocb.get_write_ptr();
-
-            // Increment the counter
-            gen_counter_incr!(ocb, $counter_name);
-
-            // Jump to the existing side exit
-            jmp_ptr(ocb, $existing_side_exit);
-
-            // Pointer to the side-exit code
-            code_ptr
-        }
-    };
+pub fn gen_counter_increment(cb: &mut CodeBlock, ptr: *const u8) {
+    // Use REG1 because there might be return value in REG0
+    mov(cb, REG1, const_ptr_opnd(ptr));
+    write_lock_prefix(cb); // for ractors.
+    add(cb, mem_opnd(64, REG1, 0), imm_opnd(1));
 }
 
 // Save the incremented PC on the CFP
 // This is necessary when callees can raise or allocate
-fn jit_save_pc(jit: &JITState, cb: &mut CodeBlock, scratch_reg: X86Opnd) {
+pub fn jit_save_pc(jit: &JITState, cb: &mut CodeBlock, scratch_reg: YJitOpnd) {
     let pc: *mut VALUE = jit.get_pc();
     let ptr: *mut VALUE = unsafe {
         let cur_insn_len = insn_len(jit.get_opcode()) as isize;
@@ -273,7 +74,7 @@ fn jit_save_pc(jit: &JITState, cb: &mut CodeBlock, scratch_reg: X86Opnd) {
 /// This realigns the interpreter SP with the JIT SP
 /// Note: this will change the current value of REG_SP,
 ///       which could invalidate memory operands
-fn gen_save_sp(cb: &mut CodeBlock, ctx: &mut Context) {
+pub fn gen_save_sp(cb: &mut CodeBlock, ctx: &mut Context) {
     if ctx.get_sp_offset() != 0 {
         let stack_pointer = ctx.sp_opnd(0);
         lea(cb, REG_SP, stack_pointer);
@@ -283,116 +84,8 @@ fn gen_save_sp(cb: &mut CodeBlock, ctx: &mut Context) {
     }
 }
 
-/// jit_save_pc() + gen_save_sp(). Should be used before calling a routine that
-/// could:
-///  - Perform GC allocation
-///  - Take the VM lock through RB_VM_LOCK_ENTER()
-///  - Perform Ruby method call
-fn jit_prepare_routine_call(
-    jit: &mut JITState,
-    ctx: &mut Context,
-    cb: &mut CodeBlock,
-    scratch_reg: X86Opnd,
-) {
-    jit.record_boundary_patch_point = true;
-    jit_save_pc(jit, cb, scratch_reg);
-    gen_save_sp(cb, ctx);
-
-    // In case the routine calls Ruby methods, it can set local variables
-    // through Kernel#binding and other means.
-    ctx.clear_local_types();
-}
-
-/// Record the current codeblock write position for rewriting into a jump into
-/// the outlined block later. Used to implement global code invalidation.
-fn record_global_inval_patch(cb: &mut CodeBlock, outline_block_target_pos: CodePtr) {
-    CodegenGlobals::push_global_inval_patch(cb.get_write_ptr(), outline_block_target_pos);
-}
-
-/// Verify the ctx's types and mappings against the compile-time stack, self,
-/// and locals.
-fn verify_ctx(jit: &JITState, ctx: &Context) {
-    fn obj_info_str<'a>(val: VALUE) -> &'a str {
-        unsafe { CStr::from_ptr(rb_obj_info(val)).to_str().unwrap() }
-    }
-
-    // Only able to check types when at current insn
-    assert!(jit_at_current_insn(jit));
-
-    let self_val = jit_peek_at_self(jit);
-    let self_val_type = Type::from(self_val);
-
-    // Verify self operand type
-    if self_val_type.diff(ctx.get_opnd_type(SelfOpnd)) == usize::MAX {
-        panic!(
-            "verify_ctx: ctx self type ({:?}) incompatible with actual value of self {}",
-            ctx.get_opnd_type(SelfOpnd),
-            obj_info_str(self_val)
-        );
-    }
-
-    // Verify stack operand types
-    let top_idx = cmp::min(ctx.get_stack_size(), MAX_TEMP_TYPES as u16);
-    for i in 0..top_idx {
-        let (learned_mapping, learned_type) = ctx.get_opnd_mapping(StackOpnd(i));
-        let stack_val = jit_peek_at_stack(jit, ctx, i as isize);
-        let val_type = Type::from(stack_val);
-
-        match learned_mapping {
-            TempMapping::MapToSelf => {
-                if self_val != stack_val {
-                    panic!(
-                        "verify_ctx: stack value was mapped to self, but values did not match!\n  stack: {}\n  self: {}",
-                        obj_info_str(stack_val),
-                        obj_info_str(self_val)
-                    );
-                }
-            }
-            TempMapping::MapToLocal(local_idx) => {
-                let local_val = jit_peek_at_local(jit, local_idx.into());
-                if local_val != stack_val {
-                    panic!(
-                        "verify_ctx: stack value was mapped to local, but values did not match\n  stack: {}\n  local {}: {}",
-                        obj_info_str(stack_val),
-                        local_idx,
-                        obj_info_str(local_val)
-                    );
-                }
-            }
-            TempMapping::MapToStack => {}
-        }
-
-        // If the actual type differs from the learned type
-        if val_type.diff(learned_type) == usize::MAX {
-            panic!(
-                "verify_ctx: ctx type ({:?}) incompatible with actual value on stack: {}",
-                learned_type,
-                obj_info_str(stack_val)
-            );
-        }
-    }
-
-    // Verify local variable types
-    let local_table_size = unsafe { get_iseq_body_local_table_size(jit.iseq) };
-    let top_idx: usize = cmp::min(local_table_size as usize, MAX_TEMP_TYPES);
-    for i in 0..top_idx {
-        let learned_type = ctx.get_local_type(i);
-        let local_val = jit_peek_at_local(jit, i as i32);
-        let local_type = Type::from(local_val);
-
-        if local_type.diff(learned_type) == usize::MAX {
-            panic!(
-                "verify_ctx: ctx type ({:?}) incompatible with actual value of local: {} (type {:?})",
-                learned_type,
-                obj_info_str(local_val),
-                local_type
-            );
-        }
-    }
-}
-
 /// Generate an exit to return to the interpreter
-fn gen_exit(exit_pc: *mut VALUE, ctx: &Context, cb: &mut CodeBlock) -> CodePtr {
+pub fn gen_exit(exit_pc: *mut VALUE, ctx: &Context, cb: &mut CodeBlock) -> CodePtr {
     let code_ptr = cb.get_write_ptr();
 
     add_comment(cb, "exit to interpreter");
@@ -430,7 +123,7 @@ fn gen_exit(exit_pc: *mut VALUE, ctx: &Context, cb: &mut CodeBlock) -> CodePtr {
 // to the interpreter when it cannot service a stub by generating new code.
 // Before coming here, branch_stub_hit() takes care of fully reconstructing
 // interpreter state.
-fn gen_code_for_exit_from_stub(ocb: &mut OutlinedCb) -> CodePtr {
+pub fn gen_code_for_exit_from_stub(ocb: &mut OutlinedCb) -> CodePtr {
     let ocb = ocb.unwrap();
     let code_ptr = ocb.get_write_ptr();
 
@@ -444,51 +137,6 @@ fn gen_code_for_exit_from_stub(ocb: &mut OutlinedCb) -> CodePtr {
     ret(ocb);
 
     return code_ptr;
-}
-
-// :side-exit:
-// Get an exit for the current instruction in the outlined block. The code
-// for each instruction often begins with several guards before proceeding
-// to do work. When guards fail, an option we have is to exit to the
-// interpreter at an instruction boundary. The piece of code that takes
-// care of reconstructing interpreter state and exiting out of generated
-// code is called the side exit.
-//
-// No guards change the logic for reconstructing interpreter state at the
-// moment, so there is one unique side exit for each context. Note that
-// it's incorrect to jump to the side exit after any ctx stack push/pop operations
-// since they change the logic required for reconstructing interpreter state.
-fn get_side_exit(jit: &mut JITState, ocb: &mut OutlinedCb, ctx: &Context) -> CodePtr {
-    match jit.side_exit_for_pc {
-        None => {
-            let exit_code = gen_exit(jit.pc, ctx, ocb.unwrap());
-            jit.side_exit_for_pc = Some(exit_code);
-            exit_code
-        }
-        Some(code_ptr) => code_ptr,
-    }
-}
-
-// Ensure that there is an exit for the start of the block being compiled.
-// Block invalidation uses this exit.
-pub fn jit_ensure_block_entry_exit(jit: &mut JITState, ocb: &mut OutlinedCb) {
-    let blockref = jit.block.clone();
-    let mut block = blockref.borrow_mut();
-    let block_ctx = block.get_ctx();
-    let blockid = block.get_blockid();
-
-    if block.entry_exit.is_some() {
-        return;
-    }
-
-    if jit.insn_idx == blockid.idx {
-        // We are compiling the first instruction in the block.
-        // Generate the exit with the cache in jitstate.
-        block.entry_exit = Some(get_side_exit(jit, ocb, &block_ctx));
-    } else {
-        let pc = unsafe { rb_iseq_pc_at_idx(blockid.iseq, blockid.idx) };
-        block.entry_exit = Some(gen_exit(pc, &block_ctx, ocb.unwrap()));
-    }
 }
 
 // Generate a runtime guard that ensures the PC is at the expected
@@ -525,7 +173,7 @@ fn gen_pc_guard(cb: &mut CodeBlock, iseq: IseqPtr, insn_idx: u32) {
 }
 
 // Landing code for when c_return tracing is enabled. See full_cfunc_return().
-fn gen_full_cfunc_return(ocb: &mut OutlinedCb) -> CodePtr {
+pub fn gen_full_cfunc_return(ocb: &mut OutlinedCb) -> CodePtr {
     let cb = ocb.unwrap();
     let code_ptr = cb.get_write_ptr();
 
@@ -553,7 +201,7 @@ fn gen_full_cfunc_return(ocb: &mut OutlinedCb) -> CodePtr {
 
 /// Generate a continuation for leave that exits to the interpreter at REG_CFP->pc.
 /// This is used by gen_leave() and gen_entry_prologue()
-fn gen_leave_exit(ocb: &mut OutlinedCb) -> CodePtr {
+pub fn gen_leave_exit(ocb: &mut OutlinedCb) -> CodePtr {
     let ocb = ocb.unwrap();
     let code_ptr = ocb.get_write_ptr();
 
@@ -645,196 +293,6 @@ fn gen_check_ints(cb: &mut CodeBlock, side_exit: CodePtr) {
     jnz_ptr(cb, side_exit);
 }
 
-// Generate a stubbed unconditional jump to the next bytecode instruction.
-// Blocks that are part of a guard chain can use this to share the same successor.
-fn jump_to_next_insn(
-    jit: &mut JITState,
-    current_context: &Context,
-    cb: &mut CodeBlock,
-    ocb: &mut OutlinedCb,
-) {
-    // Reset the depth since in current usages we only ever jump to to
-    // chain_depth > 0 from the same instruction.
-    let mut reset_depth = *current_context;
-    reset_depth.reset_chain_depth();
-
-    let jump_block = BlockId {
-        iseq: jit.iseq,
-        idx: jit_next_insn_idx(jit),
-    };
-
-    // We are at the end of the current instruction. Record the boundary.
-    if jit.record_boundary_patch_point {
-        let next_insn = unsafe { jit.pc.offset(insn_len(jit.opcode).try_into().unwrap()) };
-        let exit_pos = gen_exit(next_insn, &reset_depth, ocb.unwrap());
-        record_global_inval_patch(cb, exit_pos);
-        jit.record_boundary_patch_point = false;
-    }
-
-    // Generate the jump instruction
-    gen_direct_jump(jit, &reset_depth, jump_block, cb);
-}
-
-// Compile a sequence of bytecode instructions for a given basic block version.
-// Part of gen_block_version().
-// Note: this function will mutate its context while generating code,
-//       but the input start_ctx argument should remain immutable.
-pub fn gen_single_block(
-    blockid: BlockId,
-    start_ctx: &Context,
-    ec: EcPtr,
-    cb: &mut CodeBlock,
-    ocb: &mut OutlinedCb,
-) -> Result<BlockRef, ()> {
-    // Limit the number of specialized versions for this block
-    let mut ctx = limit_block_versions(blockid, start_ctx);
-
-    verify_blockid(blockid);
-    assert!(!(blockid.idx == 0 && ctx.get_stack_size() > 0));
-
-    // Instruction sequence to compile
-    let iseq = blockid.iseq;
-    let iseq_size = unsafe { get_iseq_encoded_size(iseq) };
-    let mut insn_idx: c_uint = blockid.idx;
-    let starting_insn_idx = insn_idx;
-
-    // Allocate the new block
-    let blockref = Block::new(blockid, &ctx);
-
-    // Initialize a JIT state object
-    let mut jit = JITState::new(&blockref);
-    jit.iseq = blockid.iseq;
-    jit.ec = Some(ec);
-
-    // Mark the start position of the block
-    blockref.borrow_mut().set_start_addr(cb.get_write_ptr());
-
-    // For each instruction to compile
-    // NOTE: could rewrite this loop with a std::iter::Iterator
-    while insn_idx < iseq_size {
-        // Get the current pc and opcode
-        let pc = unsafe { rb_iseq_pc_at_idx(iseq, insn_idx) };
-        // try_into() call below is unfortunate. Maybe pick i32 instead of usize for opcodes.
-        let opcode: usize = unsafe { rb_iseq_opcode_at_pc(iseq, pc) }
-            .try_into()
-            .unwrap();
-
-        // opt_getinlinecache wants to be in a block all on its own. Cut the block short
-        // if we run into it. See gen_opt_getinlinecache() for details.
-        if opcode == OP_OPT_GETINLINECACHE && insn_idx > starting_insn_idx {
-            jump_to_next_insn(&mut jit, &ctx, cb, ocb);
-            break;
-        }
-
-        // Set the current instruction
-        jit.insn_idx = insn_idx;
-        jit.opcode = opcode;
-        jit.pc = pc;
-        jit.side_exit_for_pc = None;
-
-        // If previous instruction requested to record the boundary
-        if jit.record_boundary_patch_point {
-            // Generate an exit to this instruction and record it
-            let exit_pos = gen_exit(jit.pc, &ctx, ocb.unwrap());
-            record_global_inval_patch(cb, exit_pos);
-            jit.record_boundary_patch_point = false;
-        }
-
-        // In debug mode, verify our existing assumption
-        if cfg!(debug_assertions) && get_option!(verify_ctx) && jit_at_current_insn(&jit) {
-            verify_ctx(&jit, &ctx);
-        }
-
-        // Lookup the codegen function for this instruction
-        let mut status = CantCompile;
-        if let Some(gen_fn) = get_gen_fn(VALUE(opcode)) {
-            // :count-placement:
-            // Count bytecode instructions that execute in generated code.
-            // Note that the increment happens even when the output takes side exit.
-            gen_counter_incr!(cb, exec_instruction);
-
-            // Add a comment for the name of the YARV instruction
-            add_comment(cb, &insn_name(opcode));
-
-            // If requested, dump instructions for debugging
-            if get_option!(dump_insns) {
-                println!("compiling {}", insn_name(opcode));
-                print_str(cb, &format!("executing {}", insn_name(opcode)));
-            }
-
-            // Call the code generation function
-            status = gen_fn(&mut jit, &mut ctx, cb, ocb);
-        }
-
-        // If we can't compile this instruction
-        // exit to the interpreter and stop compiling
-        if status == CantCompile {
-            let mut block = jit.block.borrow_mut();
-
-            // TODO: if the codegen function makes changes to ctx and then return YJIT_CANT_COMPILE,
-            // the exit this generates would be wrong. We could save a copy of the entry context
-            // and assert that ctx is the same here.
-            let exit = gen_exit(jit.pc, &ctx, cb);
-
-            // If this is the first instruction in the block, then we can use
-            // the exit for block->entry_exit.
-            if insn_idx == block.get_blockid().idx {
-                block.entry_exit = Some(exit);
-            }
-
-            break;
-        }
-
-        // For now, reset the chain depth after each instruction as only the
-        // first instruction in the block can concern itself with the depth.
-        ctx.reset_chain_depth();
-
-        // Move to the next instruction to compile
-        insn_idx += insn_len(opcode);
-
-        // If the instruction terminates this block
-        if status == EndBlock {
-            break;
-        }
-    }
-
-    // Finish filling out the block
-    {
-        let mut block = jit.block.borrow_mut();
-
-        // Mark the end position of the block
-        block.set_end_addr(cb.get_write_ptr());
-
-        // Store the index of the last instruction in the block
-        block.set_end_idx(insn_idx);
-    }
-
-    // We currently can't handle cases where the request is for a block that
-    // doesn't go to the next instruction.
-    //assert!(!jit.record_boundary_patch_point);
-
-    // If code for the block doesn't fit, fail
-    if cb.has_dropped_bytes() || ocb.unwrap().has_dropped_bytes() {
-        return Err(());
-    }
-
-    // TODO: we may want a feature for this called dump_insns? Can leave commented for now
-    /*
-    if (YJIT_DUMP_MODE >= 2) {
-        // Dump list of compiled instrutions
-        fprintf(stderr, "Compiled the following for iseq=%p:\n", (void *)iseq);
-        for (uint32_t idx = block->blockid.idx; idx < insn_idx; ) {
-            int opcode = yjit_opcode_at_pc(iseq, yjit_iseq_pc_at_idx(iseq, idx));
-            fprintf(stderr, "  %04d %s\n", idx, insn_name(opcode));
-            idx += insn_len(opcode);
-        }
-    }
-    */
-
-    // Block compiled successfully
-    Ok(blockref)
-}
-
 fn gen_nop(
     _jit: &mut JITState,
     _ctx: &mut Context,
@@ -887,17 +345,17 @@ fn gen_dupn(
         return CantCompile;
     }
 
-    let opnd1: X86Opnd = ctx.stack_opnd(1);
-    let opnd0: X86Opnd = ctx.stack_opnd(0);
+    let opnd1: YJitOpnd = ctx.stack_opnd(1);
+    let opnd0: YJitOpnd = ctx.stack_opnd(0);
 
     let mapping1 = ctx.get_opnd_mapping(StackOpnd(1));
     let mapping0 = ctx.get_opnd_mapping(StackOpnd(0));
 
-    let dst1: X86Opnd = ctx.stack_push_mapping(mapping1);
+    let dst1: YJitOpnd = ctx.stack_push_mapping(mapping1);
     mov(cb, REG0, opnd1);
     mov(cb, dst1, REG0);
 
-    let dst0: X86Opnd = ctx.stack_push_mapping(mapping0);
+    let dst0: YJitOpnd = ctx.stack_push_mapping(mapping0);
     mov(cb, REG0, opnd0);
     mov(cb, dst0, REG0);
 
@@ -920,8 +378,8 @@ fn stack_swap(
     cb: &mut CodeBlock,
     offset0: u16,
     offset1: u16,
-    _reg0: X86Opnd,
-    _reg1: X86Opnd,
+    _reg0: YJitOpnd,
+    _reg1: YJitOpnd,
 ) {
     let opnd0 = ctx.stack_opnd(offset0 as i32);
     let opnd1 = ctx.stack_opnd(offset1 as i32);
@@ -1014,7 +472,7 @@ fn gen_putself(
     mov(cb, REG0, cf_opnd);
 
     // Write it on the stack
-    let stack_top: X86Opnd = ctx.stack_push_self();
+    let stack_top: YJitOpnd = ctx.stack_push_self();
     mov(cb, stack_top, REG0);
 
     KeepCompiling
@@ -1029,7 +487,7 @@ fn gen_putspecialobject(
     let object_type = jit_get_arg(jit, 0);
 
     if object_type == VALUE(VM_SPECIAL_OBJECT_VMCORE) {
-        let stack_top: X86Opnd = ctx.stack_push(Type::UnknownHeap);
+        let stack_top: YJitOpnd = ctx.stack_push(Type::UnknownHeap);
         jit_mov_gc_ptr(jit, cb, REG0, unsafe { rb_mRubyVMFrozenCore });
         mov(cb, stack_top, REG0);
         KeepCompiling
@@ -1050,8 +508,8 @@ fn gen_setn(
     let nval: VALUE = jit_get_arg(jit, 0);
     let VALUE(n) = nval;
 
-    let top_val: X86Opnd = ctx.stack_pop(0);
-    let dst_opnd: X86Opnd = ctx.stack_opnd(n.try_into().unwrap());
+    let top_val: YJitOpnd = ctx.stack_pop(0);
+    let dst_opnd: YJitOpnd = ctx.stack_opnd(n.try_into().unwrap());
     mov(cb, REG0, top_val);
     mov(cb, dst_opnd, REG0);
 
@@ -1267,7 +725,7 @@ fn gen_newrange(
 
 fn guard_object_is_heap(
     cb: &mut CodeBlock,
-    object_opnd: X86Opnd,
+    object_opnd: YJitOpnd,
     _ctx: &mut Context,
     side_exit: CodePtr,
 ) {
@@ -1284,8 +742,8 @@ fn guard_object_is_heap(
 
 fn guard_object_is_array(
     cb: &mut CodeBlock,
-    object_opnd: X86Opnd,
-    flags_opnd: X86Opnd,
+    object_opnd: YJitOpnd,
+    flags_opnd: YJitOpnd,
     _ctx: &mut Context,
     side_exit: CodePtr,
 ) {
@@ -1467,7 +925,7 @@ fn slot_to_local_idx(iseq: IseqPtr, slot_idx: i32) -> u32 {
 }
 
 // Get EP at level from CFP
-fn gen_get_ep(cb: &mut CodeBlock, reg: X86Opnd, level: u32) {
+fn gen_get_ep(cb: &mut CodeBlock, reg: YJitOpnd, level: u32) {
     // Load environment pointer EP from CFP
     let ep_opnd = mem_opnd(64, REG_CFP, RUBY_OFFSET_CFP_EP);
     mov(cb, reg, ep_opnd);
@@ -2295,7 +1753,7 @@ fn guard_two_fixnums(ctx: &mut Context, cb: &mut CodeBlock, side_exit: CodePtr) 
 }
 
 // Conditional move operation used by comparison operators
-type CmovFn = fn(cb: &mut CodeBlock, opnd0: X86Opnd, opnd1: X86Opnd) -> ();
+type CmovFn = fn(cb: &mut CodeBlock, opnd0: YJitOpnd, opnd1: YJitOpnd) -> ();
 
 fn gen_fixnum_cmp(
     jit: &mut JITState,
@@ -3480,7 +2938,7 @@ fn jit_protected_callee_ancestry_guard(
 // Codegen for rb_obj_not().
 // Note, caller is responsible for generating all the right guards, including
 // arity guards.
-fn jit_rb_obj_not(
+pub fn jit_rb_obj_not(
     _jit: &mut JITState,
     ctx: &mut Context,
     cb: &mut CodeBlock,
@@ -3514,7 +2972,7 @@ fn jit_rb_obj_not(
 }
 
 // Codegen for rb_true()
-fn jit_rb_true(
+pub fn jit_rb_true(
     _jit: &mut JITState,
     ctx: &mut Context,
     cb: &mut CodeBlock,
@@ -3533,7 +2991,7 @@ fn jit_rb_true(
 }
 
 // Codegen for rb_false()
-fn jit_rb_false(
+pub fn jit_rb_false(
     _jit: &mut JITState,
     ctx: &mut Context,
     cb: &mut CodeBlock,
@@ -3553,7 +3011,7 @@ fn jit_rb_false(
 
 // Codegen for rb_obj_equal()
 // object identity comparison
-fn jit_rb_obj_equal(
+pub fn jit_rb_obj_equal(
     _jit: &mut JITState,
     ctx: &mut Context,
     cb: &mut CodeBlock,
@@ -3579,7 +3037,7 @@ fn jit_rb_obj_equal(
     true
 }
 
-fn jit_rb_str_bytesize(
+pub fn jit_rb_str_bytesize(
     _jit: &mut JITState,
     ctx: &mut Context,
     cb: &mut CodeBlock,
@@ -3606,7 +3064,7 @@ fn jit_rb_str_bytesize(
 // When String#to_s is called on a String instance, the method returns self and
 // most of the overhead comes from setting up the method call. We observed that
 // this situation happens a lot in some workloads.
-fn jit_rb_str_to_s(
+pub fn jit_rb_str_to_s(
     _jit: &mut JITState,
     _ctx: &mut Context,
     cb: &mut CodeBlock,
@@ -3626,7 +3084,7 @@ fn jit_rb_str_to_s(
     false
 }
 
-fn jit_thread_s_current(
+pub fn jit_thread_s_current(
     _jit: &mut JITState,
     ctx: &mut Context,
     cb: &mut CodeBlock,
@@ -5584,7 +5042,7 @@ fn gen_opt_invokebuiltin_delegate(
 }
 
 /// Maps a YARV opcode to a code generation function (if supported)
-fn get_gen_fn(opcode: VALUE) -> Option<InsnGenFn> {
+pub fn get_gen_fn(opcode: VALUE) -> Option<InsnGenFn> {
     let VALUE(opcode) = opcode;
     assert!(opcode < VM_INSTRUCTION_SIZE);
 
@@ -5679,148 +5137,9 @@ fn get_gen_fn(opcode: VALUE) -> Option<InsnGenFn> {
     }
 }
 
-// Return true when the codegen function generates code.
-// known_recv_klass is non-NULL when the caller has used jit_guard_known_klass().
-// See yjit_reg_method().
-type MethodGenFn = fn(
-    jit: &mut JITState,
-    ctx: &mut Context,
-    cb: &mut CodeBlock,
-    ocb: &mut OutlinedCb,
-    ci: *const rb_callinfo,
-    cme: *const rb_callable_method_entry_t,
-    block: Option<IseqPtr>,
-    argc: i32,
-    known_recv_class: *const VALUE,
-) -> bool;
-
-/// Global state needed for code generation
-pub struct CodegenGlobals {
-    /// Inline code block (fast path)
-    inline_cb: CodeBlock,
-
-    /// Outlined code block (slow path)
-    outlined_cb: OutlinedCb,
-
-    /// Code for exiting back to the interpreter from the leave instruction
-    leave_exit_code: CodePtr,
-
-    // For exiting from YJIT frame from branch_stub_hit().
-    // Filled by gen_code_for_exit_from_stub().
-    stub_exit_code: CodePtr,
-
-    // Code for full logic of returning from C method and exiting to the interpreter
-    outline_full_cfunc_return_pos: CodePtr,
-
-    /// For implementing global code invalidation
-    global_inval_patches: Vec<CodepagePatch>,
-
-    /// For implementing global code invalidation. The number of bytes counting from the beginning
-    /// of the inline code block that should not be changed. After patching for global invalidation,
-    /// no one should make changes to the invalidated code region anymore. This is used to
-    /// break out of invalidation race when there are multiple ractors.
-    inline_frozen_bytes: usize,
-
-    // Methods for generating code for hardcoded (usually C) methods
-    method_codegen_table: HashMap<u64, MethodGenFn>,
-}
-
-/// For implementing global code invalidation. A position in the inline
-/// codeblock to patch into a JMP rel32 which jumps into some code in
-/// the outlined codeblock to exit to the interpreter.
-pub struct CodepagePatch {
-    pub inline_patch_pos: CodePtr,
-    pub outlined_target_pos: CodePtr,
-}
-
-/// Private singleton instance of the codegen globals
-static mut CODEGEN_GLOBALS: Option<CodegenGlobals> = None;
-
 impl CodegenGlobals {
-    /// Initialize the codegen globals
-    pub fn init() {
-        // Executable memory size in MiB
-        let mem_size = get_option!(exec_mem_size) * 1024 * 1024;
-
-        #[cfg(not(test))]
-        let (mut cb, mut ocb) = {
-            let page_size = unsafe { rb_yjit_get_page_size() }.as_usize();
-            let mem_block: *mut u8 = unsafe { alloc_exec_mem(mem_size.try_into().unwrap()) };
-            let cb = CodeBlock::new(mem_block, mem_size / 2, page_size);
-            let ocb = OutlinedCb::wrap(CodeBlock::new(
-                unsafe { mem_block.add(mem_size / 2) },
-                mem_size / 2,
-                page_size,
-            ));
-            (cb, ocb)
-        };
-
-        // In test mode we're not linking with the C code
-        // so we don't allocate executable memory
-        #[cfg(test)]
-        let mut cb = CodeBlock::new_dummy(mem_size / 2);
-        #[cfg(test)]
-        let mut ocb = OutlinedCb::wrap(CodeBlock::new_dummy(mem_size / 2));
-
-        let leave_exit_code = gen_leave_exit(&mut ocb);
-
-        let stub_exit_code = gen_code_for_exit_from_stub(&mut ocb);
-
-        // Generate full exit code for C func
-        let cfunc_exit_code = gen_full_cfunc_return(&mut ocb);
-
-        // Mark all code memory as executable
-        cb.mark_all_executable();
-        ocb.unwrap().mark_all_executable();
-
-        let mut codegen_globals = CodegenGlobals {
-            inline_cb: cb,
-            outlined_cb: ocb,
-            leave_exit_code: leave_exit_code,
-            stub_exit_code: stub_exit_code,
-            outline_full_cfunc_return_pos: cfunc_exit_code,
-            global_inval_patches: Vec::new(),
-            inline_frozen_bytes: 0,
-            method_codegen_table: HashMap::new(),
-        };
-
-        // Register the method codegen functions
-        codegen_globals.reg_method_codegen_fns();
-
-        // Initialize the codegen globals instance
-        unsafe {
-            CODEGEN_GLOBALS = Some(codegen_globals);
-        }
-    }
-
-    // Register a specialized codegen function for a particular method. Note that
-    // the if the function returns true, the code it generates runs without a
-    // control frame and without interrupt checks. To avoid creating observable
-    // behavior changes, the codegen function should only target simple code paths
-    // that do not allocate and do not make method calls.
-    fn yjit_reg_method(&mut self, klass: VALUE, mid_str: &str, gen_fn: MethodGenFn) {
-        let id_string = std::ffi::CString::new(mid_str).expect("couldn't convert to CString!");
-        let mid = unsafe { rb_intern(id_string.as_ptr()) };
-        let me = unsafe { rb_method_entry_at(klass, mid) };
-
-        if me.is_null() {
-            panic!("undefined optimized method!");
-        }
-
-        // For now, only cfuncs are supported
-        //RUBY_ASSERT(me && me->def);
-        //RUBY_ASSERT(me->def->type == VM_METHOD_TYPE_CFUNC);
-
-        let method_serial = unsafe {
-            let def = (*me).def;
-            get_def_method_serial(def)
-        };
-
-        self.method_codegen_table.insert(method_serial, gen_fn);
-    }
-
     /// Register codegen functions for some Ruby core methods
-    fn reg_method_codegen_fns(&mut self) {
+    pub fn reg_method_codegen_fns(&mut self) {
         unsafe {
             // Specialization for C methods. See yjit_reg_method() for details.
             self.yjit_reg_method(rb_cBasicObject, "!", jit_rb_obj_not);
@@ -5848,67 +5167,18 @@ impl CodegenGlobals {
             );
         }
     }
+}
 
-    /// Get a mutable reference to the codegen globals instance
-    pub fn get_instance() -> &'static mut CodegenGlobals {
-        unsafe { CODEGEN_GLOBALS.as_mut().unwrap() }
-    }
+pub fn gen_call_branch_stub_hit(ocb: &mut CodeBlock, target_idx: u32, branch_ptr: *const u8, branch_stub_hit_ptr: *mut u8) {
+    // Call branch_stub_hit(branch_idx, target_idx, ec)
+    mov(ocb, C_ARG_REGS[2], REG_EC);
+    mov(ocb, C_ARG_REGS[1], uimm_opnd(target_idx as u64));
+    mov(ocb, C_ARG_REGS[0], const_ptr_opnd(branch_ptr));
+    call_ptr(ocb, REG0, branch_stub_hit_ptr);
 
-    /// Get a mutable reference to the inline code block
-    pub fn get_inline_cb() -> &'static mut CodeBlock {
-        &mut CodegenGlobals::get_instance().inline_cb
-    }
-
-    /// Get a mutable reference to the outlined code block
-    pub fn get_outlined_cb() -> &'static mut OutlinedCb {
-        &mut CodegenGlobals::get_instance().outlined_cb
-    }
-
-    pub fn get_leave_exit_code() -> CodePtr {
-        CodegenGlobals::get_instance().leave_exit_code
-    }
-
-    pub fn get_stub_exit_code() -> CodePtr {
-        CodegenGlobals::get_instance().stub_exit_code
-    }
-
-    pub fn push_global_inval_patch(i_pos: CodePtr, o_pos: CodePtr) {
-        let patch = CodepagePatch {
-            inline_patch_pos: i_pos,
-            outlined_target_pos: o_pos,
-        };
-        CodegenGlobals::get_instance()
-            .global_inval_patches
-            .push(patch);
-    }
-
-    // Drain the list of patches and return it
-    pub fn take_global_inval_patches() -> Vec<CodepagePatch> {
-        let globals = CodegenGlobals::get_instance();
-        mem::take(&mut globals.global_inval_patches)
-    }
-
-    pub fn get_inline_frozen_bytes() -> usize {
-        CodegenGlobals::get_instance().inline_frozen_bytes
-    }
-
-    pub fn set_inline_frozen_bytes(frozen_bytes: usize) {
-        CodegenGlobals::get_instance().inline_frozen_bytes = frozen_bytes;
-    }
-
-    pub fn get_outline_full_cfunc_return_pos() -> CodePtr {
-        CodegenGlobals::get_instance().outline_full_cfunc_return_pos
-    }
-
-    pub fn look_up_codegen_method(method_serial: u64) -> Option<MethodGenFn> {
-        let table = &CodegenGlobals::get_instance().method_codegen_table;
-
-        let option_ref = table.get(&method_serial);
-        match option_ref {
-            None => None,
-            Some(&mgf) => Some(mgf), // Deref
-        }
-    }
+    // Jump to the address returned by the
+    // branch_stub_hit call
+    jmp_rm(ocb, RAX);
 }
 
 #[cfg(test)]
@@ -5942,13 +5212,6 @@ mod tests {
         let (_, ctx, mut cb, _) = setup_codegen();
         gen_exit(0 as *mut VALUE, &ctx, &mut cb);
         assert!(cb.get_write_pos() > 0);
-    }
-
-    #[test]
-    fn test_get_side_exit() {
-        let (mut jit, ctx, _, mut ocb) = setup_codegen();
-        get_side_exit(&mut jit, &mut ocb, &ctx);
-        assert!(ocb.unwrap().get_write_pos() > 0);
     }
 
     #[test]
