+++ yjit/src/cruby.rs
@@ -601,7 +601,7 @@ impl From<VALUE> for i32 {
 /// Produce a Ruby string from a Rust string slice
 #[cfg(feature = "asm_comments")]
 pub fn rust_str_to_ruby(str: &str) -> VALUE {
-    unsafe { rb_utf8_str_new(str.as_ptr() as *const i8, str.len() as i64) }
+    unsafe { rb_utf8_str_new(str.as_ptr() as *const ::std::os::raw::c_char, str.len() as i64) }
 }
 
 /// Produce a Ruby symbol from a Rust string slice
